saved downloaded apps in /tmp folder and remove after installing

+++ apps-development/dbgate/uninstall.sh
@@ -1,3 +1,7 @@
-sudo apt purge -y --autoremove dbgate
+# uninstall only if it's already installed
+if apt search dbgate | grep installed &> /dev/null
+then
+  sudo apt purge -y --autoremove dbgate
+fi
 
 sudo rm -rf ~/.dbgate
